+++ src/mega/mega.py
@@ -1,5 +1,7 @@
 import re
 import json
+import secrets
+import hashlib
 from Crypto.Cipher import AES
 from Crypto.PublicKey import RSA
 from Crypto.Util import Counter
@@ -41,10 +43,27 @@ class Mega(object):
         return self
 
     def _login_user(self, email, password):
-        password_aes = prepare_key(str_to_a32(password))
-        uh = stringhash(email, password_aes)
-        resp = self._api_request({'a': 'us', 'user': email, 'uh': uh})
-        # if numeric error code response
+        email = email.lower()
+        get_user_salt_resp = self._api_request({'a': 'us0', 'user': email})
+        user_salt = None
+        try:
+            user_salt = base64_to_a32(get_user_salt_resp['s'])
+        except KeyError:
+            # v1 user account
+            password_aes = prepare_key(str_to_a32(password))
+            user_hash = stringhash(email, password_aes)
+        else:
+            # v2 user account
+            pbkdf2_key = hashlib.pbkdf2_hmac(
+                hash_name='sha512',
+                password=password.encode(),
+                salt=a32_to_str(user_salt),
+                iterations=100000,
+                dklen=32
+            )
+            password_aes = str_to_a32(pbkdf2_key[:16])
+            user_hash = base64_url_encode(pbkdf2_key[-16:])
+        resp = self._api_request({'a': 'us', 'user': email, 'uh': user_hash})
         if isinstance(resp, int):
             raise RequestError(resp)
         self._login_process(resp, password_aes)
@@ -70,7 +89,6 @@ class Mega(object):
         )
 
         resp = self._api_request({'a': 'us', 'user': user})
-        # if numeric error code response
         if isinstance(resp, int):
             raise RequestError(resp)
         self._login_process(resp, password_key)
@@ -113,7 +131,6 @@ class Mega(object):
                     self.rsa_private_key[1]
                 )
             )
-
             sid = '%x' % rsa_decrypter.key._decrypt(encrypted_sid)
             sid = binascii.unhexlify('0' + sid if len(sid) % 2 else sid)
             self.sid = base64_url_encode(sid[:43])
@@ -135,20 +152,8 @@ class Mega(object):
             params=params,
             data=json.dumps(data),
             timeout=self.timeout,
-            headers={
-                'Origin':
-                'https://mega.nz',
-                'Referer':
-                'https://mega.nz/login',
-                'User-Agent': (
-                    'Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:69.0) '
-                    'Gecko/20100101 Firefox/69.0'
-                ),
-            }
         )
         json_resp = json.loads(req.text)
-
-        # if numeric error code response
         if isinstance(json_resp, int):
             raise RequestError(json_resp)
         return json_resp[0]
@@ -163,9 +168,6 @@ class Mega(object):
             raise RequestError('Url key missing')
 
     def _process_file(self, file, shared_keys):
-        """
-        Process a file
-        """
         if file['t'] == 0 or file['t'] == 1:
             keys = dict(
                 keypart.split(':', 1)
@@ -194,6 +196,13 @@ class Mega(object):
                         key = keys[hkey]
                         key = decrypt_key(base64_to_a32(key), shared_key)
                         break
+            if file['h'] and file['h'] in shared_keys.get('EXP', ()):
+                shared_key = shared_keys['EXP'][file['h']]
+                encrypted_key = str_to_a32(
+                    base64_url_decode(file['k'].split(':')[-1])
+                )
+                key = decrypt_key(encrypted_key, shared_key)
+                file['shared_folder_key'] = shared_key
             if key is not None:
                 # file
                 if file['t'] == 0:
@@ -244,9 +253,7 @@ class Mega(object):
                 shared_keys[s_item['u']] = {}
             if s_item['h'] in ok_dict:
                 shared_keys[s_item['u']][s_item['h']] = ok_dict[s_item['h']]
-
-    ##########################################################################
-    # GET
+        self.shared_keys = shared_keys
 
     def find_path_descriptor(self, path):
         """
@@ -264,8 +271,11 @@ class Mega(object):
         for foldername in paths:
             if foldername != '':
                 for file in files.items():
-                    if file[1]['a'] and file[1]['t'] and \
-                            file[1]['a']['n'] == foldername:
+                    if (
+                        file[1]['a'] and
+                        file[1]['t'] and
+                        file[1]['a']['n'] == foldername
+                    ):
                         if parent_desc == file[1]['p']:
                             parent_desc = file[0]
                             found = True
@@ -275,22 +285,40 @@ class Mega(object):
                     return None
         return parent_desc
 
-    def find(self, filename):
+    def find(self, filename=None, handle=None):
         """
         Return file object from given filename
         """
+        from pathlib import Path
+        path = Path(filename)
+        filename = path.name
         files = self.get_files()
+        parent_dir_name = path.parent.name
         for file in list(files.items()):
-            if not isinstance(file[1]['a'], dict):
-                continue
-            if file[1]['a'] and file[1]['a']['n'] == filename:
+            parent_node_id = None
+            if parent_dir_name:
+                parent_node_id = self.find_path_descriptor(parent_dir_name)
+                if (
+                    filename and parent_node_id and
+                    file[1]['a'] and file[1]['a']['n'] == filename
+                    and parent_node_id == file[1]['p']
+                ):
+                    return file
+            # if not isinstance(file[1]['a'], dict):
+            #     continue
+            if (
+                filename and
+                file[1]['a'] and file[1]['a']['n'] == filename
+            ):
+                return file
+            if handle and file[1]['h'] == handle:
                 return file
 
     def get_files(self):
         """
         Get all files in account
         """
-        files = self._api_request({'a': 'f', 'c': 1})
+        files = self._api_request({'a': 'f', 'c': 1, 'r': 1})
         files_dict = {}
         shared_keys = {}
         self._init_shared_keys(files, shared_keys)
@@ -336,6 +364,31 @@ class Mega(object):
                 )
             decrypted_key = a32_to_base64(file['key'])
             return '{0}://{1}/#!{2}!{3}'.format(
+                self.schema, self.domain, public_handle, decrypted_key
+            )
+        else:
+            raise ValidationError('File id and key must be present')
+
+    def _node_data(self, node):
+        try:
+            return node[1]
+        except (IndexError, KeyError):
+            return node
+
+    def get_folder_link(self, file):
+        try:
+            file = file[1]
+        except (IndexError, KeyError):
+            pass
+        if 'h' in file and 'k' in file:
+            public_handle = self._api_request({'a': 'l', 'n': file['h']})
+            if public_handle == -11:
+                raise RequestError(
+                    "Can't get a public link from that file "
+                    "(is this a shared file?)"
+                )
+            decrypted_key = a32_to_base64(file['shared_folder_key'])
+            return '{0}://{1}/#F!{2}!{3}'.format(
                 self.schema, self.domain, public_handle, decrypted_key
             )
         else:
@@ -442,8 +495,6 @@ class Mega(object):
         if 'balance' in user_data:
             return user_data['balance']
 
-    ##########################################################################
-    # DELETE
     def delete(self, public_handle):
         """
         Delete a file by its public handle
@@ -491,8 +542,6 @@ class Mega(object):
                 post_list.append({"a": "d", "n": file, "i": self.request_id})
             return self._api_request(post_list)
 
-    ##########################################################################
-    # DOWNLOAD
     def download(self, file, dest_path=None, dest_filename=None):
         """
         Download a file by it's file object
@@ -505,6 +554,68 @@ class Mega(object):
             dest_filename=dest_filename,
             is_public=False
         )
+
+    def _export_file(self, node):
+        self._api_request([
+            {
+                'a': 'l',
+                'n': node[1]['h'],
+                'i': self.request_id
+            }
+        ])
+        return self.get_link(node)
+
+    def export(self, path):
+        self.get_files()
+        folder = self.find(path)
+        if folder[1]['t'] == 0:
+            return self._export_file(folder)
+        if folder:
+            try:
+                # If already exported
+                return self.get_folder_link(folder)
+            except (RequestError, KeyError):
+                pass
+
+        user_id = folder[1]['u']
+        user_pub_key = self._api_request({'a': 'uk', 'u': user_id})['pubk']
+        node_key = folder[1]['k']
+
+        master_key_cipher = AES.new(a32_to_str(self.master_key), AES.MODE_ECB)
+        ha = base64_url_encode(
+            master_key_cipher.encrypt(folder[1]['h'] + folder[1]['h'])
+        )
+
+        share_key = secrets.token_bytes(16)
+        ok = base64_url_encode(master_key_cipher.encrypt(share_key))
+
+        share_key_cipher = AES.new(share_key, AES.MODE_ECB)
+        encrypted_node_key = base64_url_encode(
+            share_key_cipher.encrypt(a32_to_str(node_key))
+        )
+
+        node_id = folder[1]['h']
+        request_body = [
+            {
+            'a': 's2',
+            'n': node_id,
+            's': [{
+                'u': 'EXP',
+                'r': 0
+            }],
+            'i': self.request_id,
+            'ok': ok,
+            'ha': ha,
+            'cr': [
+                [node_id],
+                [node_id],
+                [0, 0, encrypted_node_key]
+            ]
+        }]
+        self._api_request(request_body)
+        nodes = self.get_files()
+        link = self.get_folder_link(nodes[node_id])
+        return link
 
     def download_url(self, url, dest_path=None, dest_filename=None):
         """
@@ -631,8 +742,6 @@ class Mega(object):
 
         shutil.move(temp_output_file.name, dest_path + file_name)
 
-    ##########################################################################
-    # UPLOAD
     def upload(self, filename, dest=None, dest_filename=None):
         # determine storage node
         if dest is None:
@@ -767,10 +876,8 @@ class Mega(object):
         # update attributes
         data = self._api_request(
             {
-                'a':
-                'p',
-                't':
-                dest,
+                'a':'p',
+                't': dest,
                 'n': [
                     {
                         'h': 'xxxxxxxx',
@@ -779,8 +886,7 @@ class Mega(object):
                         'k': encrypted_key
                     }
                 ],
-                'i':
-                self.request_id
+                'i': self.request_id
             }
         )
         return data
@@ -881,21 +987,6 @@ class Mega(object):
                 }
             )
 
-    def get_contacts(self):
-        raise NotImplementedError()
-        # TODO implement this
-        # sn param below = maxaction var with function getsc() in mega.co.nz js
-        # seens to be the 'sn' attrib of the previous request response...
-        # requests goto /sc rather than
-
-        # req = requests.post(
-        # '{0}://g.api.{1}/sc'.format(self.schema, self.domain),
-        #     params={'sn': 'ZMxcQ_DmHnM', 'ssl': '1'},
-        #     data=json.dumps(None),
-        #     timeout=self.timeout)
-        # json_resp = json.loads(req.text)
-        # print json_resp
-
     def get_public_url_info(self, url):
         """
         Get size and name from a public url, dict returned
@@ -917,8 +1008,6 @@ class Mega(object):
         Get size and name of a public file
         """
         data = self._api_request({'a': 'g', 'p': file_handle, 'ssm': 1})
-
-        # if numeric error code response
         if isinstance(data, int):
             raise RequestError(data)