+++ HISTORY.rst
@@ -6,7 +6,9 @@ Release History
 0.9.21 (unreleased)
 +++++++++++++++++++
 
-- Nothing changed yet.
+- Removes broken method ``get_contacts()``.
+- Adds support for login with a v2 Mega user account.
+- Adds ``export()`` method to share a file or folder, returning public share URL with key.
 
 
 0.9.20 (2019-10-17)